requirements.txt
Django
djangorestframework
twilio
django-phonenumber-field[phonenumbers]
python-decouple
drf-social-oauth2